Fix tests first pass
