+++ tests/common/util/MockAmendmentBlockHandler.hpp
@@ -19,11 +19,11 @@
 
 #pragma once
 
-#include "etlng/AmendmentBlockHandlerInterface.hpp"
+#include "etl/AmendmentBlockHandlerInterface.hpp"
 
 #include <gmock/gmock.h>
 
-struct MockAmendmentBlockHandler : etlng::AmendmentBlockHandlerInterface {
+struct MockAmendmentBlockHandler : etl::AmendmentBlockHandlerInterface {
     MOCK_METHOD(void, notifyAmendmentBlocked, (), (override));
     MOCK_METHOD(void, stop, (), (override));
 };
