+++ tests/common/util/MockLoadBalancer.hpp
@@ -20,8 +20,8 @@
 #pragma once
 
 #include "etl/ETLState.hpp"
-#include "etlng/InitialLoadObserverInterface.hpp"
-#include "etlng/LoadBalancerInterface.hpp"
+#include "etl/InitialLoadObserverInterface.hpp"
+#include "etl/LoadBalancerInterface.hpp"
 #include "rpc/Errors.hpp"
 #include "util/FakeFetchResponse.hpp"
 
@@ -38,13 +38,13 @@
 #include <string>
 #include <vector>
 
-struct MockNgLoadBalancer : etlng::LoadBalancerInterface {
+struct MockNgLoadBalancer : etl::LoadBalancerInterface {
     using RawLedgerObjectType = FakeLedgerObject;
 
     MOCK_METHOD(
-        etlng::InitialLedgerLoadResult,
+        etl::InitialLedgerLoadResult,
         loadInitialLedger,
-        (uint32_t, etlng::InitialLoadObserverInterface&, std::chrono::steady_clock::duration),
+        (uint32_t, etl::InitialLoadObserverInterface&, std::chrono::steady_clock::duration),
         (override)
     );
     MOCK_METHOD(
