+++ tests/unit/etl/LoadBalancerTests.cpp
@@ -1,7 +1,7 @@
 //------------------------------------------------------------------------------
 /*
     This file is part of clio: https://github.com/XRPLF/clio
-    Copyright (c) 2024, the clio developers.
+    Copyright (c) 2025, the clio developers.
 
     Permission to use, copy, modify, and distribute this software for any
     purpose with or without fee is hereby granted, provided that the above
@@ -17,7 +17,10 @@
 */
 //==============================================================================
 
+#include "etl/InitialLoadObserverInterface.hpp"
 #include "etl/LoadBalancer.hpp"
+#include "etl/LoadBalancerInterface.hpp"
+#include "etl/Models.hpp"
 #include "etl/Source.hpp"
 #include "rpc/Errors.hpp"
 #include "util/AsioContextTestFixture.hpp"
@@ -25,7 +28,7 @@
 #include "util/MockNetworkValidatedLedgers.hpp"
 #include "util/MockPrometheus.hpp"
 #include "util/MockRandomGenerator.hpp"
-#include "util/MockSource.hpp"
+#include "util/MockSourceNg.hpp"
 #include "util/MockSubscriptionManager.hpp"
 #include "util/NameGenerator.hpp"
 #include "util/config/Array.hpp"
@@ -62,7 +65,9 @@ using namespace util::config;
 using testing::Return;
 using namespace util::prometheus;
 
-static constexpr auto kTWO_SOURCES_LEDGER_RESPONSE = R"JSON({
+namespace {
+
+constinit auto const kTWO_SOURCES_LEDGER_RESPONSE = R"JSON({
     "etl_sources": [
         {
             "ip": "127.0.0.1",
@@ -77,7 +82,7 @@ static constexpr auto kTWO_SOURCES_LEDGER_RESPONSE = R"JSON({
     ]
 })JSON";
 
-static constexpr auto kTHREE_SOURCES_LEDGER_RESPONSE = R"JSON({
+constinit auto const kTHREE_SOURCES_LEDGER_RESPONSE = R"JSON({
     "etl_sources": [
         {
             "ip": "127.0.0.1",
@@ -97,7 +102,7 @@ static constexpr auto kTHREE_SOURCES_LEDGER_RESPONSE = R"JSON({
     ]
 })JSON";
 
-inline static ClioConfigDefinition
+inline ClioConfigDefinition
 getParseLoadBalancerConfig(boost::json::value val)
 {
     ClioConfigDefinition config{
@@ -118,7 +123,24 @@ getParseLoadBalancerConfig(boost::json::value val)
     return config;
 }
 
-struct LoadBalancerConstructorTests : util::prometheus::WithPrometheus, MockBackendTestStrict {
+struct InitialLoadObserverMock : etl::InitialLoadObserverInterface {
+    MOCK_METHOD(
+        void,
+        onInitialLoadGotMoreObjects,
+        (uint32_t, std::vector<etl::model::Object> const&, std::optional<std::string>),
+        (override)
+    );
+
+    void
+    onInitialLoadGotMoreObjects(uint32_t seq, std::vector<etl::model::Object> const& data)
+    {
+        onInitialLoadGotMoreObjects(seq, data, std::nullopt);
+    }
+};
+
+}  // namespace
+
+struct LoadBalancerConstructorNgTests : util::prometheus::WithPrometheus, MockBackendTestStrict {
     std::unique_ptr<LoadBalancer>
     makeLoadBalancer()
     {
@@ -142,12 +164,12 @@ protected:
     MockRandomGenerator* randomGenerator_ = nullptr;
     StrictMockSubscriptionManagerSharedPtr subscriptionManager_;
     StrictMockNetworkValidatedLedgersPtr networkManager_;
-    StrictMockSourceFactory sourceFactory_{2};
+    StrictMockSourceNgFactory sourceFactory_{2};
     boost::asio::io_context ioContext_;
     boost::json::value configJson_ = boost::json::parse(kTWO_SOURCES_LEDGER_RESPONSE);
 };
 
-TEST_F(LoadBalancerConstructorTests, construct)
+TEST_F(LoadBalancerConstructorNgTests, construct)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled).WillOnce(Return(boost::json::object{}));
@@ -157,14 +179,13 @@ TEST_F(LoadBalancerConstructorTests, construct)
     makeLoadBalancer();
 }
 
-TEST_F(LoadBalancerConstructorTests, forwardingTimeoutPassedToSourceFactory)
+TEST_F(LoadBalancerConstructorNgTests, forwardingTimeoutPassedToSourceFactory)
 {
     auto const forwardingTimeout = 10;
     configJson_.as_object()["forwarding"] = boost::json::object{{"cache_timeout", float{forwardingTimeout}}};
     EXPECT_CALL(
         sourceFactory_,
         makeSource(
-            testing::_,
             testing::_,
             testing::_,
             testing::_,
@@ -183,7 +204,7 @@ TEST_F(LoadBalancerConstructorTests, forwardingTimeoutPassedToSourceFactory)
     makeLoadBalancer();
 }
 
-TEST_F(LoadBalancerConstructorTests, fetchETLState_AllSourcesFail)
+TEST_F(LoadBalancerConstructorNgTests, fetchETLState_AllSourcesFail)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled)
@@ -193,7 +214,7 @@ TEST_F(LoadBalancerConstructorTests, fetchETLState_AllSourcesFail)
     EXPECT_THROW({ makeLoadBalancer(); }, std::logic_error);
 }
 
-TEST_F(LoadBalancerConstructorTests, fetchETLState_AllSourcesReturnError)
+TEST_F(LoadBalancerConstructorNgTests, fetchETLState_AllSourcesReturnError)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled)
@@ -203,7 +224,7 @@ TEST_F(LoadBalancerConstructorTests, fetchETLState_AllSourcesReturnError)
     EXPECT_THROW({ makeLoadBalancer(); }, std::logic_error);
 }
 
-TEST_F(LoadBalancerConstructorTests, fetchETLState_Source1Fails0OK)
+TEST_F(LoadBalancerConstructorNgTests, fetchETLState_Source1Fails0OK)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled).WillOnce(Return(boost::json::object{}));
@@ -214,7 +235,7 @@ TEST_F(LoadBalancerConstructorTests, fetchETLState_Source1Fails0OK)
     makeLoadBalancer();
 }
 
-TEST_F(LoadBalancerConstructorTests, fetchETLState_Source0Fails1OK)
+TEST_F(LoadBalancerConstructorNgTests, fetchETLState_Source0Fails1OK)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled)
@@ -225,7 +246,7 @@ TEST_F(LoadBalancerConstructorTests, fetchETLState_Source0Fails1OK)
     makeLoadBalancer();
 }
 
-TEST_F(LoadBalancerConstructorTests, fetchETLState_DifferentNetworkID)
+TEST_F(LoadBalancerConstructorNgTests, fetchETLState_DifferentNetworkID)
 {
     auto const source1Json = boost::json::parse(R"JSON({"result": {"info": {"network_id": 0}}})JSON");
     auto const source2Json = boost::json::parse(R"JSON({"result": {"info": {"network_id": 1}}})JSON");
@@ -236,7 +257,7 @@ TEST_F(LoadBalancerConstructorTests, fetchETLState_DifferentNetworkID)
     EXPECT_THROW({ makeLoadBalancer(); }, std::logic_error);
 }
 
-TEST_F(LoadBalancerConstructorTests, fetchETLState_AllSourcesFailButAllowNoEtlIsTrue)
+TEST_F(LoadBalancerConstructorNgTests, fetchETLState_AllSourcesFailButAllowNoEtlIsTrue)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled).WillOnce(Return(boost::json::object{}));
@@ -249,7 +270,7 @@ TEST_F(LoadBalancerConstructorTests, fetchETLState_AllSourcesFailButAllowNoEtlIs
     makeLoadBalancer();
 }
 
-TEST_F(LoadBalancerConstructorTests, fetchETLState_DifferentNetworkIDButAllowNoEtlIsTrue)
+TEST_F(LoadBalancerConstructorNgTests, fetchETLState_DifferentNetworkIDButAllowNoEtlIsTrue)
 {
     auto const source1Json = boost::json::parse(R"JSON({"result": {"info": {"network_id": 0}}})JSON");
     auto const source2Json = boost::json::parse(R"JSON({"result": {"info": {"network_id": 1}}})JSON");
@@ -263,8 +284,8 @@ TEST_F(LoadBalancerConstructorTests, fetchETLState_DifferentNetworkIDButAllowNoE
     makeLoadBalancer();
 }
 
-struct LoadBalancerOnConnectHookTests : LoadBalancerConstructorTests {
-    LoadBalancerOnConnectHookTests()
+struct LoadBalancerOnConnectHookNgTests : LoadBalancerConstructorNgTests {
+    LoadBalancerOnConnectHookNgTests()
     {
         EXPECT_CALL(sourceFactory_, makeSource).Times(2);
         EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled).WillOnce(Return(boost::json::object{}));
@@ -278,7 +299,7 @@ protected:
     std::unique_ptr<LoadBalancer> loadBalancer_;
 };
 
-TEST_F(LoadBalancerOnConnectHookTests, sourcesConnect)
+TEST_F(LoadBalancerOnConnectHookNgTests, sourcesConnect)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(true));
     EXPECT_CALL(sourceFactory_.sourceAt(0), setForwarding(true));
@@ -287,7 +308,7 @@ TEST_F(LoadBalancerOnConnectHookTests, sourcesConnect)
     sourceFactory_.callbacksAt(1).onConnect();
 }
 
-TEST_F(LoadBalancerOnConnectHookTests, sourcesConnect_Source0IsNotConnected)
+TEST_F(LoadBalancerOnConnectHookNgTests, sourcesConnect_Source0IsNotConnected)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(false));
     EXPECT_CALL(sourceFactory_.sourceAt(0), setForwarding(false));
@@ -305,7 +326,7 @@ TEST_F(LoadBalancerOnConnectHookTests, sourcesConnect_Source0IsNotConnected)
     sourceFactory_.callbacksAt(0).onConnect();
 }
 
-TEST_F(LoadBalancerOnConnectHookTests, sourcesConnect_BothSourcesAreNotConnected)
+TEST_F(LoadBalancerOnConnectHookNgTests, sourcesConnect_BothSourcesAreNotConnected)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(false));
     EXPECT_CALL(sourceFactory_.sourceAt(0), setForwarding(false));
@@ -326,17 +347,17 @@ TEST_F(LoadBalancerOnConnectHookTests, sourcesConnect_BothSourcesAreNotConnected
     sourceFactory_.callbacksAt(0).onConnect();
 }
 
-struct LoadBalancerStopTests : LoadBalancerOnConnectHookTests, SyncAsioContextTest {};
+struct LoadBalancerStopNgTests : LoadBalancerOnConnectHookNgTests, SyncAsioContextTest {};
 
-TEST_F(LoadBalancerStopTests, stopCallsSourcesStop)
+TEST_F(LoadBalancerStopNgTests, stopCallsSourcesStop)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), stop);
     EXPECT_CALL(sourceFactory_.sourceAt(1), stop);
     runSyncOperation([this](boost::asio::yield_context yield) { loadBalancer_->stop(yield); });
 }
 
-struct LoadBalancerOnDisconnectHookTests : LoadBalancerOnConnectHookTests {
-    LoadBalancerOnDisconnectHookTests()
+struct LoadBalancerOnDisconnectHookNgTests : LoadBalancerOnConnectHookNgTests {
+    LoadBalancerOnDisconnectHookNgTests()
     {
         EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(true));
         EXPECT_CALL(sourceFactory_.sourceAt(0), setForwarding(true));
@@ -348,7 +369,7 @@ struct LoadBalancerOnDisconnectHookTests : LoadBalancerOnConnectHookTests {
     }
 };
 
-TEST_F(LoadBalancerOnDisconnectHookTests, source0Disconnects)
+TEST_F(LoadBalancerOnDisconnectHookNgTests, source0Disconnects)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(false));
     EXPECT_CALL(sourceFactory_.sourceAt(0), setForwarding(false));
@@ -357,12 +378,12 @@ TEST_F(LoadBalancerOnDisconnectHookTests, source0Disconnects)
     sourceFactory_.callbacksAt(0).onDisconnect(true);
 }
 
-TEST_F(LoadBalancerOnDisconnectHookTests, source1Disconnects)
+TEST_F(LoadBalancerOnDisconnectHookNgTests, source1Disconnects)
 {
     sourceFactory_.callbacksAt(1).onDisconnect(false);
 }
 
-TEST_F(LoadBalancerOnDisconnectHookTests, source0DisconnectsAndConnectsBack)
+TEST_F(LoadBalancerOnDisconnectHookNgTests, source0DisconnectsAndConnectsBack)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(false));
     EXPECT_CALL(sourceFactory_.sourceAt(0), setForwarding(false));
@@ -373,13 +394,13 @@ TEST_F(LoadBalancerOnDisconnectHookTests, source0DisconnectsAndConnectsBack)
     sourceFactory_.callbacksAt(0).onConnect();
 }
 
-TEST_F(LoadBalancerOnDisconnectHookTests, source1DisconnectsAndConnectsBack)
+TEST_F(LoadBalancerOnDisconnectHookNgTests, source1DisconnectsAndConnectsBack)
 {
     sourceFactory_.callbacksAt(1).onDisconnect(false);
     sourceFactory_.callbacksAt(1).onConnect();
 }
 
-TEST_F(LoadBalancerOnConnectHookTests, bothSourcesDisconnectAndConnectBack)
+TEST_F(LoadBalancerOnConnectHookNgTests, bothSourcesDisconnectAndConnectBack)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(false));
     EXPECT_CALL(sourceFactory_.sourceAt(0), setForwarding(false));
@@ -396,8 +417,8 @@ TEST_F(LoadBalancerOnConnectHookTests, bothSourcesDisconnectAndConnectBack)
     sourceFactory_.callbacksAt(1).onConnect();
 }
 
-struct LoadBalancer3SourcesTests : LoadBalancerConstructorTests {
-    LoadBalancer3SourcesTests()
+struct LoadBalancer3SourcesNgTests : LoadBalancerConstructorNgTests {
+    LoadBalancer3SourcesNgTests()
     {
         sourceFactory_.setSourcesNumber(3);
         configJson_ = boost::json::parse(kTHREE_SOURCES_LEDGER_RESPONSE);
@@ -416,7 +437,7 @@ protected:
     std::unique_ptr<LoadBalancer> loadBalancer_;
 };
 
-TEST_F(LoadBalancer3SourcesTests, forwardingUpdate)
+TEST_F(LoadBalancer3SourcesNgTests, forwardingUpdate)
 {
     // Source 2 is connected first
     EXPECT_CALL(sourceFactory_.sourceAt(0), isConnected()).WillOnce(Return(false));
@@ -435,58 +456,64 @@ TEST_F(LoadBalancer3SourcesTests, forwardingUpdate)
     sourceFactory_.callbacksAt(0).onDisconnect(false);
 }
 
-struct LoadBalancerLoadInitialLedgerTests : LoadBalancerOnConnectHookTests {
+struct LoadBalancerLoadInitialLedgerNgTests : LoadBalancerOnConnectHookNgTests {
 protected:
     uint32_t const sequence_ = 123;
     uint32_t const numMarkers_ = 16;
-    std::pair<std::vector<std::string>, bool> const response_ = {{"1", "2", "3"}, true};
+    InitialLedgerLoadResult const response_{std::vector<std::string>{"1", "2", "3"}};
+    testing::StrictMock<InitialLoadObserverMock> observer_;
 };
 
-TEST_F(LoadBalancerLoadInitialLedgerTests, load)
+TEST_F(LoadBalancerLoadInitialLedgerNgTests, load)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), hasLedger(sequence_)).WillOnce(Return(true));
-    EXPECT_CALL(sourceFactory_.sourceAt(0), loadInitialLedger(sequence_, numMarkers_)).WillOnce(Return(response_));
+    EXPECT_CALL(sourceFactory_.sourceAt(0), loadInitialLedger(sequence_, numMarkers_, testing::_))
+        .WillOnce(Return(response_));
 
-    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_), response_.first);
+    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_, observer_, std::chrono::milliseconds{1}), response_.value());
 }
 
-TEST_F(LoadBalancerLoadInitialLedgerTests, load_source0DoesntHaveLedger)
+TEST_F(LoadBalancerLoadInitialLedgerNgTests, load_source0DoesntHaveLedger)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), hasLedger(sequence_)).WillOnce(Return(false));
     EXPECT_CALL(sourceFactory_.sourceAt(1), hasLedger(sequence_)).WillOnce(Return(true));
-    EXPECT_CALL(sourceFactory_.sourceAt(1), loadInitialLedger(sequence_, numMarkers_)).WillOnce(Return(response_));
+    EXPECT_CALL(sourceFactory_.sourceAt(1), loadInitialLedger(sequence_, numMarkers_, testing::_))
+        .WillOnce(Return(response_));
 
-    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_), response_.first);
+    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_, observer_, std::chrono::milliseconds{1}), response_.value());
 }
 
-TEST_F(LoadBalancerLoadInitialLedgerTests, load_bothSourcesDontHaveLedger)
+TEST_F(LoadBalancerLoadInitialLedgerNgTests, load_bothSourcesDontHaveLedger)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), hasLedger(sequence_)).Times(2).WillRepeatedly(Return(false));
     EXPECT_CALL(sourceFactory_.sourceAt(1), hasLedger(sequence_)).WillOnce(Return(false)).WillOnce(Return(true));
-    EXPECT_CALL(sourceFactory_.sourceAt(1), loadInitialLedger(sequence_, numMarkers_)).WillOnce(Return(response_));
+    EXPECT_CALL(sourceFactory_.sourceAt(1), loadInitialLedger(sequence_, numMarkers_, testing::_))
+        .WillOnce(Return(response_));
 
-    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_, std::chrono::milliseconds{1}), response_.first);
+    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_, observer_, std::chrono::milliseconds{1}), response_.value());
 }
 
-TEST_F(LoadBalancerLoadInitialLedgerTests, load_source0ReturnsStatusFalse)
+TEST_F(LoadBalancerLoadInitialLedgerNgTests, load_source0ReturnsStatusFalse)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), hasLedger(sequence_)).WillOnce(Return(true));
-    EXPECT_CALL(sourceFactory_.sourceAt(0), loadInitialLedger(sequence_, numMarkers_))
-        .WillOnce(Return(std::make_pair(std::vector<std::string>{}, false)));
+    EXPECT_CALL(sourceFactory_.sourceAt(0), loadInitialLedger(sequence_, numMarkers_, testing::_))
+        .WillOnce(Return(std::unexpected{InitialLedgerLoadError::Errored}));
     EXPECT_CALL(sourceFactory_.sourceAt(1), hasLedger(sequence_)).WillOnce(Return(true));
-    EXPECT_CALL(sourceFactory_.sourceAt(1), loadInitialLedger(sequence_, numMarkers_)).WillOnce(Return(response_));
+    EXPECT_CALL(sourceFactory_.sourceAt(1), loadInitialLedger(sequence_, numMarkers_, testing::_))
+        .WillOnce(Return(response_));
 
-    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_), response_.first);
+    EXPECT_EQ(loadBalancer_->loadInitialLedger(sequence_, observer_, std::chrono::milliseconds{1}), response_.value());
 }
 
-struct LoadBalancerLoadInitialLedgerCustomNumMarkersTests : LoadBalancerConstructorTests {
+struct LoadBalancerLoadInitialLedgerCustomNumMarkersNgTests : LoadBalancerConstructorNgTests {
 protected:
     uint32_t const numMarkers_ = 16;
     uint32_t const sequence_ = 123;
-    std::pair<std::vector<std::string>, bool> const response_ = {{"1", "2", "3"}, true};
+    InitialLedgerLoadResult const response_{std::vector<std::string>{"1", "2", "3"}};
+    testing::StrictMock<InitialLoadObserverMock> observer_;
 };
 
-TEST_F(LoadBalancerLoadInitialLedgerCustomNumMarkersTests, loadInitialLedger)
+TEST_F(LoadBalancerLoadInitialLedgerCustomNumMarkersNgTests, loadInitialLedger)
 {
     configJson_.as_object()["num_markers"] = numMarkers_;
 
@@ -498,13 +525,14 @@ TEST_F(LoadBalancerLoadInitialLedgerCustomNumMarkersTests, loadInitialLedger)
     auto loadBalancer = makeLoadBalancer();
 
     EXPECT_CALL(sourceFactory_.sourceAt(0), hasLedger(sequence_)).WillOnce(Return(true));
-    EXPECT_CALL(sourceFactory_.sourceAt(0), loadInitialLedger(sequence_, numMarkers_)).WillOnce(Return(response_));
+    EXPECT_CALL(sourceFactory_.sourceAt(0), loadInitialLedger(sequence_, numMarkers_, testing::_))
+        .WillOnce(Return(response_));
 
-    EXPECT_EQ(loadBalancer->loadInitialLedger(sequence_), response_.first);
+    EXPECT_EQ(loadBalancer->loadInitialLedger(sequence_, observer_, std::chrono::milliseconds{1}), response_.value());
 }
 
-struct LoadBalancerFetchLegerTests : LoadBalancerOnConnectHookTests {
-    LoadBalancerFetchLegerTests()
+struct LoadBalancerFetchLegerNgTests : LoadBalancerOnConnectHookNgTests {
+    LoadBalancerFetchLegerNgTests()
     {
         response_.second.set_validated(true);
     }
@@ -517,7 +545,7 @@ protected:
         std::make_pair(grpc::Status::OK, org::xrpl::rpc::v1::GetLedgerResponse{});
 };
 
-TEST_F(LoadBalancerFetchLegerTests, fetch)
+TEST_F(LoadBalancerFetchLegerNgTests, fetch)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), hasLedger(sequence_)).WillOnce(Return(true));
     EXPECT_CALL(sourceFactory_.sourceAt(0), fetchLedger(sequence_, getObjects_, getObjectNeighbors_))
@@ -526,7 +554,7 @@ TEST_F(LoadBalancerFetchLegerTests, fetch)
     EXPECT_TRUE(loadBalancer_->fetchLedger(sequence_, getObjects_, getObjectNeighbors_).has_value());
 }
 
-TEST_F(LoadBalancerFetchLegerTests, fetch_Source0ReturnsBadStatus)
+TEST_F(LoadBalancerFetchLegerNgTests, fetch_Source0ReturnsBadStatus)
 {
     auto source0Response = response_;
     source0Response.first = grpc::Status::CANCELLED;
@@ -542,7 +570,7 @@ TEST_F(LoadBalancerFetchLegerTests, fetch_Source0ReturnsBadStatus)
     EXPECT_TRUE(loadBalancer_->fetchLedger(sequence_, getObjects_, getObjectNeighbors_).has_value());
 }
 
-TEST_F(LoadBalancerFetchLegerTests, fetch_Source0ReturnsNotValidated)
+TEST_F(LoadBalancerFetchLegerNgTests, fetch_Source0ReturnsNotValidated)
 {
     auto source0Response = response_;
     source0Response.second.set_validated(false);
@@ -558,7 +586,7 @@ TEST_F(LoadBalancerFetchLegerTests, fetch_Source0ReturnsNotValidated)
     EXPECT_TRUE(loadBalancer_->fetchLedger(sequence_, getObjects_, getObjectNeighbors_).has_value());
 }
 
-TEST_F(LoadBalancerFetchLegerTests, fetch_bothSourcesFail)
+TEST_F(LoadBalancerFetchLegerNgTests, fetch_bothSourcesFail)
 {
     auto badResponse = response_;
     badResponse.second.set_validated(false);
@@ -576,8 +604,8 @@ TEST_F(LoadBalancerFetchLegerTests, fetch_bothSourcesFail)
                     .has_value());
 }
 
-struct LoadBalancerForwardToRippledTests : LoadBalancerConstructorTests, SyncAsioContextTest {
-    LoadBalancerForwardToRippledTests()
+struct LoadBalancerForwardToRippledNgTests : LoadBalancerConstructorNgTests, SyncAsioContextTest {
+    LoadBalancerForwardToRippledNgTests()
     {
         EXPECT_CALL(sourceFactory_.sourceAt(0), forwardToRippled).WillOnce(Return(boost::json::object{}));
         EXPECT_CALL(sourceFactory_.sourceAt(0), run);
@@ -591,7 +619,7 @@ protected:
     boost::json::object const response_{{"response", "other_value"}};
 };
 
-TEST_F(LoadBalancerForwardToRippledTests, forward)
+TEST_F(LoadBalancerForwardToRippledNgTests, forward)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     auto loadBalancer = makeLoadBalancer();
@@ -606,7 +634,7 @@ TEST_F(LoadBalancerForwardToRippledTests, forward)
     });
 }
 
-TEST_F(LoadBalancerForwardToRippledTests, forwardWithXUserHeader)
+TEST_F(LoadBalancerForwardToRippledNgTests, forwardWithXUserHeader)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     auto loadBalancer = makeLoadBalancer();
@@ -621,7 +649,7 @@ TEST_F(LoadBalancerForwardToRippledTests, forwardWithXUserHeader)
     });
 }
 
-TEST_F(LoadBalancerForwardToRippledTests, source0Fails)
+TEST_F(LoadBalancerForwardToRippledNgTests, source0Fails)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     auto loadBalancer = makeLoadBalancer();
@@ -641,9 +669,9 @@ TEST_F(LoadBalancerForwardToRippledTests, source0Fails)
     });
 }
 
-struct LoadBalancerForwardToRippledPrometheusTests : LoadBalancerForwardToRippledTests, WithMockPrometheus {};
+struct LoadBalancerForwardToRippledPrometheusNgTests : LoadBalancerForwardToRippledNgTests, WithMockPrometheus {};
 
-TEST_F(LoadBalancerForwardToRippledPrometheusTests, forwardingCacheEnabled)
+TEST_F(LoadBalancerForwardToRippledPrometheusNgTests, forwardingCacheEnabled)
 {
     configJson_.as_object()["forwarding"] = boost::json::object{{"cache_timeout", 10.}};
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
@@ -674,7 +702,7 @@ TEST_F(LoadBalancerForwardToRippledPrometheusTests, forwardingCacheEnabled)
     });
 }
 
-TEST_F(LoadBalancerForwardToRippledPrometheusTests, source0Fails)
+TEST_F(LoadBalancerForwardToRippledPrometheusNgTests, source0Fails)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     auto loadBalancer = makeLoadBalancer();
@@ -706,46 +734,46 @@ TEST_F(LoadBalancerForwardToRippledPrometheusTests, source0Fails)
     });
 }
 
-struct LoadBalancerForwardToRippledErrorTestBundle {
+struct LoadBalancerForwardToRippledErrorNgTestBundle {
     std::string testName;
     rpc::ClioError firstSourceError;
     rpc::ClioError secondSourceError;
     rpc::CombinedError responseExpectedError;
 };
 
-struct LoadBalancerForwardToRippledErrorTests
-    : LoadBalancerForwardToRippledTests,
-      testing::WithParamInterface<LoadBalancerForwardToRippledErrorTestBundle> {};
+struct LoadBalancerForwardToRippledErrorNgTests
+    : LoadBalancerForwardToRippledNgTests,
+      testing::WithParamInterface<LoadBalancerForwardToRippledErrorNgTestBundle> {};
 
 INSTANTIATE_TEST_SUITE_P(
-    LoadBalancerForwardToRippledErrorTests,
-    LoadBalancerForwardToRippledErrorTests,
+    LoadBalancerForwardToRippledErrorNgTests,
+    LoadBalancerForwardToRippledErrorNgTests,
     testing::Values(
-        LoadBalancerForwardToRippledErrorTestBundle{
+        LoadBalancerForwardToRippledErrorNgTestBundle{
             "ConnectionError_RequestError",
             rpc::ClioError::EtlConnectionError,
             rpc::ClioError::EtlRequestError,
             rpc::ClioError::EtlRequestError
         },
-        LoadBalancerForwardToRippledErrorTestBundle{
+        LoadBalancerForwardToRippledErrorNgTestBundle{
             "RequestError_RequestTimeout",
             rpc::ClioError::EtlRequestError,
             rpc::ClioError::EtlRequestTimeout,
             rpc::ClioError::EtlRequestTimeout
         },
-        LoadBalancerForwardToRippledErrorTestBundle{
+        LoadBalancerForwardToRippledErrorNgTestBundle{
             "RequestTimeout_InvalidResponse",
             rpc::ClioError::EtlRequestTimeout,
             rpc::ClioError::EtlInvalidResponse,
             rpc::ClioError::EtlInvalidResponse
         },
-        LoadBalancerForwardToRippledErrorTestBundle{
+        LoadBalancerForwardToRippledErrorNgTestBundle{
             "BothRequestTimeout",
             rpc::ClioError::EtlRequestTimeout,
             rpc::ClioError::EtlRequestTimeout,
             rpc::ClioError::EtlRequestTimeout
         },
-        LoadBalancerForwardToRippledErrorTestBundle{
+        LoadBalancerForwardToRippledErrorNgTestBundle{
             "InvalidResponse_RequestError",
             rpc::ClioError::EtlInvalidResponse,
             rpc::ClioError::EtlRequestError,
@@ -755,7 +783,7 @@ INSTANTIATE_TEST_SUITE_P(
     tests::util::kNAME_GENERATOR
 );
 
-TEST_P(LoadBalancerForwardToRippledErrorTests, bothSourcesFail)
+TEST_P(LoadBalancerForwardToRippledErrorNgTests, bothSourcesFail)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     auto loadBalancer = makeLoadBalancer();
@@ -777,7 +805,7 @@ TEST_P(LoadBalancerForwardToRippledErrorTests, bothSourcesFail)
     });
 }
 
-TEST_F(LoadBalancerForwardToRippledTests, forwardingCacheEnabled)
+TEST_F(LoadBalancerForwardToRippledNgTests, forwardingCacheEnabled)
 {
     configJson_.as_object()["forwarding"] = boost::json::object{{"cache_timeout", 10.}};
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
@@ -797,14 +825,14 @@ TEST_F(LoadBalancerForwardToRippledTests, forwardingCacheEnabled)
     });
 }
 
-TEST_F(LoadBalancerForwardToRippledTests, forwardingCacheDisabledOnLedgerClosedHookCalled)
+TEST_F(LoadBalancerForwardToRippledNgTests, forwardingCacheDisabledOnLedgerClosedHookCalled)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     auto loadBalancer = makeLoadBalancer();
     EXPECT_NO_THROW(sourceFactory_.callbacksAt(0).onLedgerClosed());
 }
 
-TEST_F(LoadBalancerForwardToRippledTests, onLedgerClosedHookInvalidatesCache)
+TEST_F(LoadBalancerForwardToRippledNgTests, onLedgerClosedHookInvalidatesCache)
 {
     configJson_.as_object()["forwarding"] = boost::json::object{{"cache_timeout", 10.}};
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
@@ -813,6 +841,7 @@ TEST_F(LoadBalancerForwardToRippledTests, onLedgerClosedHookInvalidatesCache)
     auto const request = boost::json::object{{"command", "server_info"}};
 
     EXPECT_CALL(*randomGenerator_, uniform(0, 1)).WillOnce(Return(0)).WillOnce(Return(1));
+
     EXPECT_CALL(
         sourceFactory_.sourceAt(0),
         forwardToRippled(request, clientIP_, LoadBalancer::kUSER_FORWARDING_X_USER_VALUE, testing::_)
@@ -832,7 +861,7 @@ TEST_F(LoadBalancerForwardToRippledTests, onLedgerClosedHookInvalidatesCache)
     });
 }
 
-TEST_F(LoadBalancerForwardToRippledTests, commandLineMissing)
+TEST_F(LoadBalancerForwardToRippledNgTests, commandLineMissing)
 {
     EXPECT_CALL(sourceFactory_, makeSource).Times(2);
     auto loadBalancer = makeLoadBalancer();
@@ -847,9 +876,9 @@ TEST_F(LoadBalancerForwardToRippledTests, commandLineMissing)
     });
 }
 
-struct LoadBalancerToJsonTests : LoadBalancerOnConnectHookTests {};
+struct LoadBalancerToJsonNgTests : LoadBalancerOnConnectHookNgTests {};
 
-TEST_F(LoadBalancerToJsonTests, toJson)
+TEST_F(LoadBalancerToJsonNgTests, toJson)
 {
     EXPECT_CALL(sourceFactory_.sourceAt(0), toJson).WillOnce(Return(boost::json::object{{"source1", "value1"}}));
     EXPECT_CALL(sourceFactory_.sourceAt(1), toJson).WillOnce(Return(boost::json::object{{"source2", "value2"}}));
