+++ tests/unit/etl/TaskManagerTests.cpp
@@ -17,13 +17,13 @@
 */
 //==============================================================================
 
-#include "etlng/ExtractorInterface.hpp"
-#include "etlng/LoaderInterface.hpp"
-#include "etlng/Models.hpp"
-#include "etlng/MonitorInterface.hpp"
-#include "etlng/SchedulerInterface.hpp"
-#include "etlng/impl/Loading.hpp"
-#include "etlng/impl/TaskManager.hpp"
+#include "etl/ExtractorInterface.hpp"
+#include "etl/LoaderInterface.hpp"
+#include "etl/Models.hpp"
+#include "etl/MonitorInterface.hpp"
+#include "etl/SchedulerInterface.hpp"
+#include "etl/impl/Loading.hpp"
+#include "etl/impl/TaskManager.hpp"
 #include "util/BinaryTestObject.hpp"
 #include "util/TestObject.hpp"
 #include "util/async/AnyExecutionContext.hpp"
@@ -43,30 +43,30 @@
 #include <semaphore>
 #include <vector>
 
-using namespace etlng::model;
-using namespace etlng::impl;
+using namespace etl::model;
+using namespace etl::impl;
 
 namespace {
 
 constinit auto const kSEQ = 30;
 constinit auto const kLEDGER_HASH = "4BC50C9B0D8515D3EAAE1E74B29A95804346C491EE1A95BF25E4AAB854A6A652";
 
-struct MockScheduler : etlng::SchedulerInterface {
+struct MockScheduler : etl::SchedulerInterface {
     MOCK_METHOD(std::optional<Task>, next, (), (override));
 };
 
-struct MockExtractor : etlng::ExtractorInterface {
+struct MockExtractor : etl::ExtractorInterface {
     MOCK_METHOD(std::optional<LedgerData>, extractLedgerWithDiff, (uint32_t), (override));
     MOCK_METHOD(std::optional<LedgerData>, extractLedgerOnly, (uint32_t), (override));
 };
 
-struct MockLoader : etlng::LoaderInterface {
-    using ExpectedType = std::expected<void, etlng::LoaderError>;
+struct MockLoader : etl::LoaderInterface {
+    using ExpectedType = std::expected<void, etl::LoaderError>;
     MOCK_METHOD(ExpectedType, load, (LedgerData const&), (override));
     MOCK_METHOD(std::optional<ripple::LedgerHeader>, loadInitialLedger, (LedgerData const&), (override));
 };
 
-struct MockMonitor : etlng::MonitorInterface {
+struct MockMonitor : etl::MonitorInterface {
     MOCK_METHOD(void, notifySequenceLoaded, (uint32_t), (override));
     MOCK_METHOD(void, notifyWriteConflict, (uint32_t), (override));
     MOCK_METHOD(
@@ -141,7 +141,7 @@ TEST_F(TaskManagerTests, LoaderGetsDataIfNextSequenceIsExtracted)
 
     EXPECT_CALL(*mockLoaderPtr_, load(testing::_))
         .Times(kTOTAL)
-        .WillRepeatedly([&](LedgerData data) -> std::expected<void, etlng::LoaderError> {
+        .WillRepeatedly([&](LedgerData data) -> std::expected<void, etl::LoaderError> {
             loaded.push_back(data.seq);
             if (loaded.size() == kTOTAL)
                 done.release();
@@ -185,13 +185,13 @@ TEST_F(TaskManagerTests, WriteConflictHandling)
 
     // First kCONFLICT_AFTER calls succeed, then we get a write conflict
     EXPECT_CALL(*mockLoaderPtr_, load(testing::_))
-        .WillRepeatedly([&](LedgerData data) -> std::expected<void, etlng::LoaderError> {
+        .WillRepeatedly([&](LedgerData data) -> std::expected<void, etl::LoaderError> {
             loaded.push_back(data.seq);
 
             if (loaded.size() == kCONFLICT_AFTER) {
                 conflictOccurred = true;
                 done.release();
-                return std::unexpected(etlng::LoaderError::WriteConflict);
+                return std::unexpected(etl::LoaderError::WriteConflict);
             }
 
             if (loaded.size() == kTOTAL)
@@ -238,13 +238,13 @@ TEST_F(TaskManagerTests, AmendmentBlockedHandling)
         });
 
     EXPECT_CALL(*mockLoaderPtr_, load(testing::_))
-        .WillRepeatedly([&](LedgerData data) -> std::expected<void, etlng::LoaderError> {
+        .WillRepeatedly([&](LedgerData data) -> std::expected<void, etl::LoaderError> {
             loaded.push_back(data.seq);
 
             if (loaded.size() == kAMENDMENT_BLOCKED_AFTER) {
                 amendmentBlockedOccurred = true;
                 done.release();
-                return std::unexpected(etlng::LoaderError::AmendmentBlocked);
+                return std::unexpected(etl::LoaderError::AmendmentBlocked);
             }
 
             if (loaded.size() == kTOTAL)
