+++ tests/unit/etl/ext/CoreTests.cpp
@@ -17,8 +17,8 @@
 */
 //==============================================================================
 
-#include "etlng/Models.hpp"
-#include "etlng/impl/ext/Core.hpp"
+#include "etl/Models.hpp"
+#include "etl/impl/ext/Core.hpp"
 #include "util/BinaryTestObject.hpp"
 #include "util/MockBackendTestFixture.hpp"
 #include "util/MockPrometheus.hpp"
@@ -31,7 +31,7 @@
 #include <utility>
 #include <vector>
 
-using namespace etlng::impl;
+using namespace etl::impl;
 using namespace data;
 
 namespace {
@@ -48,7 +48,7 @@ createTestData()
     };
 
     auto const header = createLedgerHeader(kLEDGER_HASH, kSEQ);
-    return etlng::model::LedgerData{
+    return etl::model::LedgerData{
         .transactions = std::move(transactions),
         .objects = {},
         .successors = {},
@@ -63,7 +63,7 @@ createTestData()
 
 struct CoreExtTests : util::prometheus::WithPrometheus, MockBackendTest {
 protected:
-    etlng::impl::CoreExt ext_{backend_};
+    etl::impl::CoreExt ext_{backend_};
 };
 
 TEST_F(CoreExtTests, OnLedgerDataWritesLedgerAndTransactions)
